requirements.txt
beautifulsoup4~=4.9.1
bs4~=0.0.1
certifi==2020.12.5
future==0.18.2
lxml==4.5.2
Pillow~=7.2.0
pyTelegramBotAPI==3.6.1
python-dotenv~=0.15.0
python-telegram-bot==7.0.1
requests~=2.13.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy~=1.4.2